Move useSenseiProExtension to avoid code duplication
So we can re-use the logic in the FeaturedProductSenseiPro component too! :D
